Implemented scrolling capabilities :scroll:
In order to accommodate for the keyboard when editing a subview's
contents, the views must be embedded within a scroll view. Without this,
the views being edited may be obscured by the keyboard, therefore
providing for a very poor user experience.

The trail view has been embedded within a scroll view and all of the
constraints have been updated to support this new feature. A private
method has been created which appropriately adjusts the scroll and
content views in response to notifications from the keyboard.