Set up HealthKit for saving distance information
A HealthKitManager class was added in order to handle all of the
HealthKit store transactions. This class manages authorization and
saving data. The data consists of the distances recorded for different
trail entries.

This new distance feature is not required, trails can still be created
without distances. But these distances, if entered, will be saved into
the HealthKit store, providing a nice record of walking/running
exercise.

The next task is to adapt the application's user interface to include
distances.