Fully integrated distances into the application
Distances are now properly recorded in the application itself and are
consistent accross navigation, but they have not been tested for
persistency.

The next four steps in order are: first, determine if distance data is
persistent; second, properly integrate the application with HealthKit
(the setup is done but hasn't been tested); third, clean up the code and
the user interface; fourth, write unit tests to ensure that there are no
design flaws.